fix miscalculation under macOS

+++ skew/phash.c
@@ -82,7 +82,10 @@ void init_skew() {
 uint16_t bin_nationkey(uint64_t key, uint64_t tbl_size) {
 	long row = key / (0.2 * tbl_size);
 	long bin = row * 5;
-	long offset = key - (0.18 + row * 0.2) * tbl_size;
+	// Hyrise: cast subtrahend as long. Otherwise, we got incorrect offsets on macOS/clang.
+	// E.g., for `key = 79`, `tbl_size = 100`, we got `row = 3`, `(0.18 + row * 0.2) * tbl_size) = 78.0`,
+	//       but `offset = 0` (instead of 1).
+	long offset = key - (long)((0.18 + row * 0.2) * tbl_size);
 	assert(row < 5);
 	if (offset > 0 && 0.02 * tbl_size > 0) { 
 		offset = (4*offset) / (0.02*tbl_size);
